+++ elisp/chronometrist.el
@@ -239,27 +239,38 @@ Return new position of point."
 ;; open-literate-source:1 ends here
 
 ;; [[file:chronometrist.org::*debug logging][debug logging:1]]
+(defcustom chronometrist-debug-enable nil
+  "Whether to log debugging messages."
+  :type 'boolean
+  :group 'chronometrist)
+
 (defcustom chronometrist-debug-buffer "*chronometrist-debug*"
   "Name of buffer to log debug messages to."
   :type 'string
   :group 'chronometrist)
 
-(defun chronometrist-debug (format-string &rest args)
-  "Log a debug message using `display-warning'.
-For FORMAT-STRING and ARGS are passed to `format'."
-  (display-warning '(chronometrist)
-                   (apply #'format
-                          (concat (format-time-string "[%T] ")
-                                  format-string)
-                          args)
-                   :debug chronometrist-debug-buffer))
+(define-derived-mode chronometrist-debug-log-mode special-mode "debug-log")
+
+(defun chronometrist-debug-message (format-string &rest args)
+  "Log a debug message to `chronometrist-debug-buffer'.
+FORMAT-STRING and ARGS are passed to `format'."
+  (when chronometrist-debug-enable
+    (with-current-buffer (get-buffer-create chronometrist-debug-buffer)
+      (chronometrist-debug-log-mode)
+      (let ((inhibit-read-only t))
+        (insert
+         (apply #'format
+                (concat (format-time-string "[%T] ")
+                        format-string)
+                args)
+         "\n")))))
 ;; debug logging:1 ends here
 
 ;; [[file:chronometrist.org::*reset][reset:1]]
 (defun chronometrist-reset ()
   "Reset Chronometrist's internal state."
   (interactive)
-  (chronometrist-debug "Command: reset")
+  (chronometrist-debug-message "Command: reset")
   (chronometrist-reset-backend (chronometrist-active-backend))
   (chronometrist-refresh))
 ;; reset:1 ends here
@@ -747,7 +758,7 @@ Value must be a keyword corresponding to a key in
 ;; [[file:chronometrist.org::*switch-backend][switch-backend:1]]
 (defun chronometrist-switch-backend ()
   (interactive)
-  (chronometrist-debug "Command: switch-backend")
+  (chronometrist-debug-message "Command: switch-backend")
   (let* ((prompt (format "Switch to backend (current - %s): "
                          chronometrist-active-backend))
          (choice (chronometrist-read-backend-name prompt
@@ -941,7 +952,10 @@ backend file).")
 
 ;; [[file:chronometrist.org::*verify][verify:1]]
 (cl-defgeneric chronometrist-verify (backend)
-  "Check BACKEND for errors in data.")
+  "Check BACKEND for errors in data.
+Return nil if no errors are found.
+
+If an error is found, return (LINE-NUMBER . COLUMN-NUMBER) for file-based backends.")
 ;; verify:1 ends here
 
 ;; [[file:chronometrist.org::*on-file-path-change][on-file-path-change:1]]
@@ -1214,8 +1228,11 @@ expression first)."
            (new-rest-hash (when (and (>= new-length rest-start)
                                      (>= new-length rest-end))
                             (chronometrist-file-hash rest-start rest-end file)))
-           (new-last-hash (when (>= new-length rest-end)
-                            (chronometrist-file-hash rest-end new-length file))))
+           (new-last-hash (when (and (>= new-length rest-end)
+                                     (>= new-length file-length))
+                            (chronometrist-file-hash rest-end file-length file))))
+      ;; (chronometrist-debug-message "File indices - old rest-start: %s rest-end: %s file-length: %s new-length: %s"
+      ;;          rest-start rest-end file-length new-length)
       (cond ((and (= file-length new-length)
                   (equal rest-hash new-rest-hash)
                   (equal last-hash new-last-hash))
@@ -1308,7 +1325,7 @@ FS-EVENT is the event passed by the `filenotify' library (see `file-notify-add-w
                            (chronometrist-file-change-type backend)))
             (reset-watch-p (or (eq action 'deleted)
                                (eq action 'renamed))))
-      (chronometrist-debug "File change type %s" change)
+      (chronometrist-debug-message "File change type %s" change)
       ;; If only the last plist was changed, update hash table and
       ;; task list, otherwise clear and repopulate hash table.
       (cond ((or reset-watch-p
@@ -1393,7 +1410,7 @@ STREAM (which is the value of `current-buffer')."
 ;; [[file:chronometrist.org::*insert][insert:1]]
 (cl-defmethod chronometrist-insert ((backend chronometrist-plist-backend) plist &key (save t))
   (chronometrist-backend-run-assertions backend)
-  (chronometrist-debug "Insert plist %s" plist)
+  (chronometrist-debug-message "Insert plist %s" plist)
   (chronometrist-sexp-in-file (chronometrist-backend-file backend)
     (goto-char (point-max))
     ;; If we're adding the first s-exp in the file, don't add a
@@ -1407,7 +1424,7 @@ STREAM (which is the value of `current-buffer')."
 
 ;; [[file:chronometrist.org::*remove-last][remove-last:1]]
 (cl-defmethod chronometrist-remove-last ((backend chronometrist-plist-backend))
-  (chronometrist-debug "Remove last plist")
+  (chronometrist-debug-message "Remove last plist")
   (chronometrist-backend-run-assertions backend)
   (when (chronometrist-backend-empty-p backend)
   (error "chronometrist-remove-last has nothing to remove in %s"
@@ -1426,7 +1443,7 @@ STREAM (which is the value of `current-buffer')."
   "Reindent the current buffer.
 This is meant to be run in `chronometrist-file' when using an s-expression backend."
   (interactive)
-  (chronometrist-debug "Command: reindent-buffer")
+  (chronometrist-debug-message "Command: reindent-buffer")
   (let (expr)
     (goto-char (point-min))
     (while (setq expr (ignore-errors (read (current-buffer))))
@@ -1464,7 +1481,7 @@ This is meant to be run in `chronometrist-file' when using an s-expression backe
   "Function run when a new plist is added at the end of a
 `chronometrist-plist-backend' file."
   (with-slots (hash-table) backend
-    (-let [(new-plist &plist :name new-task) (chronometrist-latest-record backend)]
+    (-let [(new-plist &as &plist :name new-task) (chronometrist-latest-record backend)]
       (setf hash-table (chronometrist-events-update new-plist hash-table))
       (chronometrist-add-to-task-list new-task backend))))
 ;; on-add:1 ends here
@@ -1518,7 +1535,7 @@ This is meant to be run in `chronometrist-file' when using an s-expression backe
 
 ;; [[file:chronometrist.org::*replace-last][replace-last:1]]
 (cl-defmethod chronometrist-replace-last ((backend chronometrist-plist-backend) plist)
-  (chronometrist-debug "Replace last plist with %s" plist)
+  (chronometrist-debug-message "Replace last plist with %s" plist)
   (chronometrist-sexp-in-file (chronometrist-backend-file backend)
     (goto-char (chronometrist-remove-last backend))
     (funcall chronometrist-sexp-pretty-print-function plist (current-buffer))
@@ -1572,6 +1589,7 @@ This is meant to be run in `chronometrist-file' when using an s-expression backe
 ;; [[file:chronometrist.org::*insert][insert:1]]
 (cl-defmethod chronometrist-insert ((backend chronometrist-plist-group-backend) plist &key (save t))
   (cl-check-type plist chronometrist-plist)
+  (chronometrist-debug-message "Insert %S" plist)
   (chronometrist-backend-run-assertions backend)
   (if (not plist)
       (error "%s" "`chronometrist-insert' was called with an empty plist")
@@ -1590,17 +1608,30 @@ This is meant to be run in `chronometrist-file' when using an s-expression backe
               (new-plist-group-2   (when (or plist-2 insert-new-group)
                                      (list date-today (or plist-2 plist)))))
         (goto-char (point-max))
-        ;; Is this the first plist group in the file?
-        (if latest-plist-group
-            (progn
-              (chronometrist-sexp-pre-read-check (current-buffer))
-              (chronometrist-sexp-delete-list))
+        (when (not latest-plist-group)
+          ;; first record
           (while (forward-comment 1) nil))
-        (funcall chronometrist-sexp-pretty-print-function new-plist-group-1 (current-buffer))
-        (when new-plist-group-2
-          (dotimes (_ 2)
-            (default-indent-new-line))
-          (funcall chronometrist-sexp-pretty-print-function new-plist-group-2 (current-buffer)))
+        (if (and plist-1 plist-2)
+            ;; inactive, day-crossing record
+            (progn
+              (when latest-plist-group
+                ;; not the first record
+                (chronometrist-sexp-pre-read-check (current-buffer))
+                (chronometrist-sexp-delete-list))
+              (funcall chronometrist-sexp-pretty-print-function new-plist-group-1 (current-buffer))
+              (dotimes (_ 2) (default-indent-new-line))
+              (funcall chronometrist-sexp-pretty-print-function new-plist-group-2 (current-buffer)))
+          ;; active, or non-day-crossing inactive record
+          ;; insert into new group
+          (if (or (not latest-plist-group) ;; first record
+                  insert-new-group)
+              (progn
+                (default-indent-new-line)
+                (funcall chronometrist-sexp-pretty-print-function new-plist-group-2 (current-buffer)))
+            ;; insert into existing group
+            (chronometrist-sexp-pre-read-check (current-buffer))
+            (chronometrist-sexp-delete-list)
+            (funcall chronometrist-sexp-pretty-print-function new-plist-group-1 (current-buffer))))
         (when save (save-buffer))
         t))))
 ;; insert:1 ends here
@@ -1755,6 +1786,26 @@ Return value is either a list in the form
         do (chronometrist-remove-from-task-list (plist-get plist :name) backend))
       (puthash old-date nil hash-table))))
 ;; on-remove:1 ends here
+
+;; [[file:chronometrist.org::*verify][verify:1]]
+(cl-defmethod chronometrist-verify ((backend chronometrist-plist-group-backend))
+  (with-slots (file hash-table) backend
+    ;; incorrectly ordered groups check
+    (chronometrist-loop-sexp-file for group in file
+      with old-date-iso with old-date-unix
+      with new-date-iso with new-date-unix
+      ;; while (not (bobp))
+      do (setq new-date-iso  (cl-first group)
+               new-date-unix (parse-iso8601-time-string new-date-iso))
+      when (and old-date-unix
+                (time-less-p old-date-unix
+                             new-date-unix))
+      do (cl-return (format "%s appears before %s on line %s"
+                            new-date-iso old-date-iso (line-number-at-pos)))
+      else do (setq old-date-iso  new-date-iso
+                    old-date-unix new-date-unix)
+      finally return "Yay, no errors! (...that I could find 💀)")))
+;; verify:1 ends here
 
 ;; [[file:chronometrist.org::*latest-record][latest-record:1]]
 (cl-defmethod chronometrist-latest-record ((backend chronometrist-plist-group-backend))
@@ -2411,7 +2462,7 @@ If INHIBIT-HOOKS is non-nil, the hooks
 `chronometrist-before-out-functions', and
 `chronometrist-after-out-functions' will not be run."
   (interactive "P")
-  (chronometrist-debug "Command: toggle-task %s" (if inhibit-hooks "(without hooks)" ""))
+  (chronometrist-debug-message "Command: toggle-task %s" (if inhibit-hooks "(without hooks)" ""))
   (let* ((empty-file   (chronometrist-backend-empty-p (chronometrist-active-backend)))
          (nth          (when prefix (chronometrist-goto-nth-task prefix)))
          (at-point     (chronometrist-task-at-point))
@@ -2448,7 +2499,7 @@ there is no corresponding task, do nothing."
 (defun chronometrist-add-new-task ()
   "Add a new task."
   (interactive)
-  (chronometrist-debug "Command: add-new-task")
+  (chronometrist-debug-message "Command: add-new-task")
   (chronometrist-add-new-task-button nil))
 ;; add-new-task:1 ends here
 
@@ -2461,7 +2512,7 @@ INHIBIT-HOOKS is non-nil or prefix argument is supplied.
 
 Has no effect if no task is active."
   (interactive "P")
-  (chronometrist-debug "Command: restart-task")
+  (chronometrist-debug-message "Command: restart-task")
   (if (chronometrist-current-task)
       (let* ((latest (chronometrist-latest-record (chronometrist-active-backend)))
              (plist  (plist-put latest :start (chronometrist-format-time-iso8601)))
@@ -2483,7 +2534,7 @@ INHIBIT-HOOKS is non-nil or prefix argument is supplied.
 
 Has no effect if a task is active."
   (interactive "P")
-  (chronometrist-debug "Command: extend-task")
+  (chronometrist-debug-message "Command: extend-task")
   (if (chronometrist-current-task)
       (message "Cannot extend an active task - use this after clocking out.")
     (let* ((latest (chronometrist-latest-record (chronometrist-active-backend)))
@@ -2500,7 +2551,7 @@ Has no effect if a task is active."
 (defun chronometrist-discard-active ()
   "Remove active interval from the active backend."
   (interactive)
-  (chronometrist-debug "Command: discard-active")
+  (chronometrist-debug-message "Command: discard-active")
   (let ((backend (chronometrist-active-backend)))
     (if (chronometrist-current-task backend)
         (chronometrist-remove-last backend)